refactor: consolidate review components into single directory
- Rename Reviews.tsx to AllReviews.tsx for clarity
- Move reducer and filters into Reviews directory
- Consolidate Underrated, Underseen, and Overrated components
- Combine all getProps functions into a single file with named exports
- Update all imports throughout the codebase

This refactoring improves code organization by keeping all review-related
components together and reduces duplication in the getProps functions.
