+++ src/components/Reviews/Overrated.tsx
@@ -3,30 +3,30 @@ import { type JSX, useReducer } from "react";
 import type { BackdropImageProps } from "~/api/backdrops";
 import type { ReviewListItemValue } from "~/components/ReviewListItem";
 
-import { Backdrop } from "~/components/Backdrop";
+import { Backdrop, BreadcrumbLink } from "~/components/Backdrop";
 import { GroupedList } from "~/components/GroupedList";
 import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
 import { ReviewListItem } from "~/components/ReviewListItem";
-import { Filters, SortOptions } from "~/components/Reviews.Filters";
+import { Filters, SortOptions } from "~/components/Reviews/Filters";
 import {
   Actions,
   initState,
   reducer,
   type Sort,
-} from "~/components/Reviews.reducer";
+} from "~/components/Reviews/reducer";
 import { ReviewsSubNav } from "~/components/ReviewsSubNav";
 
 export type Props = {
   backdropImageProps: BackdropImageProps;
   deck: string;
-  distinctGenres: readonly string[];
-  distinctReleaseYears: readonly string[];
-  distinctReviewYears: readonly string[];
+  distinctGenres: string[];
+  distinctReleaseYears: string[];
+  distinctReviewYears: string[];
   initialSort: Sort;
   values: ReviewListItemValue[];
 };
 
-export function Reviews({
+export function Overrated({
   backdropImageProps,
   deck,
   distinctGenres,
@@ -47,7 +47,12 @@ export function Reviews({
   return (
     <ListWithFiltersLayout
       backdrop={
-        <Backdrop deck={deck} imageProps={backdropImageProps} title="Reviews" />
+        <Backdrop
+          breadcrumb={<BreadcrumbLink href="/reviews/">Reviews</BreadcrumbLink>}
+          deck={deck}
+          imageProps={backdropImageProps}
+          title="Overrated Disappointments"
+        />
       }
       filters={
         <Filters
@@ -69,6 +74,7 @@ export function Reviews({
           {(value) => <ReviewListItem key={value.imdbId} value={value} />}
         </GroupedList>
       }
+      mastGradient={false}
       sortProps={{
         currentSortValue: state.sortValue,
         onSortChange: (e) =>
@@ -78,7 +84,7 @@ export function Reviews({
           }),
         sortOptions: <SortOptions />,
       }}
-      subNav={<ReviewsSubNav active="all" />}
+      subNav={<ReviewsSubNav active="overrated" />}
       totalCount={state.filteredValues.length}
     />
   );
