+++ src/components/Reviews/Underseen.tsx
@@ -7,13 +7,13 @@ import { Backdrop, BreadcrumbLink } from "~/components/Backdrop";
 import { GroupedList } from "~/components/GroupedList";
 import { ListWithFiltersLayout } from "~/components/ListWithFiltersLayout";
 import { ReviewListItem } from "~/components/ReviewListItem";
-import { Filters, SortOptions } from "~/components/Reviews.Filters";
+import { Filters, SortOptions } from "~/components/Reviews/Filters";
 import {
   Actions,
   initState,
   reducer,
   type Sort,
-} from "~/components/Reviews.reducer";
+} from "~/components/Reviews/reducer";
 import { ReviewsSubNav } from "~/components/ReviewsSubNav";
 
 export type Props = {
