+++ src/pages/reviews/index.astro
@@ -1,6 +1,6 @@
 ---
+import { AllReviews } from "~/components/Reviews/AllReviews";
 import { getProps } from "~/components/Reviews/getProps";
-import { Reviews } from "~/components/Reviews/Reviews";
 import Layout from "~/layouts/Layout.astro";
 
 const props = await getProps();
@@ -12,5 +12,5 @@ const props = await getProps();
   openGraph={{ description: props.deck, type: "website" }}
   title="All Reviews"
 >
-  <Reviews client:load {...props} />
+  <AllReviews client:load {...props} />
 </Layout>
