+++ src/pages/reviews/overrated/index.astro
@@ -1,6 +1,6 @@
 ---
-import { getProps } from "~/components/Overrated/getProps";
-import { Overrated } from "~/components/Overrated/Overrated";
+import { getPropsForOverrated as getProps } from "~/components/Reviews/getProps";
+import { Overrated } from "~/components/Reviews/Overrated";
 import Layout from "~/layouts/Layout.astro";
 
 const props = await getProps();
