+++ src/pages/reviews/underrated/index.astro
@@ -1,6 +1,6 @@
 ---
-import { getProps } from "~/components/Underrated/getProps";
-import { Underrated } from "~/components/Underrated/Underrated";
+import { getPropsForUnderrated as getProps } from "~/components/Reviews/getProps";
+import { Underrated } from "~/components/Reviews/Underrated";
 import Layout from "~/layouts/Layout.astro";
 
 const props = await getProps();
