+++ src/pages/reviews/underseen/index.astro
@@ -1,6 +1,6 @@
 ---
-import { getProps } from "~/components/Underseen/getProps";
-import { Underseen } from "~/components/Underseen/Underseen";
+import { getPropsForUnderseen as getProps } from "~/components/Reviews/getProps";
+import { Underseen } from "~/components/Reviews/Underseen";
 import Layout from "~/layouts/Layout.astro";
 
 const props = await getProps();
